bastard-scrapper/requirements.txt
requests==2.32.3
libsql-client==0.3.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
